Added logging to Jobs and fixed a bug with sample app.
The bug was that the same job would be submitted for each button because
it was created when the OnClickListener was created, not when the button
is tapped.

To get around this I added an abstract Builder class to Job which
subclasses can choose to implement. If they do, the can use the
"new SomeSubclass.Builder()" to the attachJobBuilderToOnClick() method and
a new Job of that subclass will be created and added to the JobManager on
click.

I'M NOT CRAZY, YOU'RE CRAZY!

This is likely not going to go into the final production of JimJam, but
for now it lives on in my heart. And the Git repo.