Added Sqlite db and persistence.
Added a button to dump and print logs of persisted jobs to the sample app.

The “20 second” job and the “requires network” job will both persist. FOREVER. Until you dump the DB.

Jobs will not be removed when they complete, and they will not be restarted when the Job Manager is created yet.

At this point you can run the sample app, fire off some persistent jobs, reboot the device, and see that the jobs are persisted. We actually deserialize them to Job objects to print them.