Allow jobs to be grouped by label.
This is a big change. The PriorityThreadPoolExecutor now has a new backing
queue. This queue understands Runnables with labels. When 2 or more jobs
with the same label are added, only 1 of them will execute at the same
time.

For now I have left in the debug logging for the new classes. There is now
a minor amount of packaging, so not everything is flat! More package
sorting to come.